Quetzio Server class

Create R6-based server module to generate and hold the state of your questionnaire

Public fields

source_list

List containing the data for all the inputs

description

List containing survey instruction and description

div_id

ID of the div containing the survey UI

module_id

ID of the shiny module

button_labels

character vector of length two with labels for submit button in active and disabled state

is_done

logical reactiveVal indicating if the survey has been completed

message

reactiveVal catching any warning messages

answers

reactiveVal object containing list with answers to questions

module_ui_id

character string used to generate UI. It needs to be modified when linking the questionnaires

Methods

Public methods

quetzio_server$toggle_ui()

quetzio_server$new()

quetzio_server$get_answers_df()

quetzio_server$update_labels()

quetzio_server$clone()

Method toggle_ui()

method to change the state of the UI

Usage

quetzio_server$toggle_ui(x)

Arguments

x

logical indicating what should be the state of the UI. If TRUE, then the UI will be rendered.

Method new()

Initializing the 'quetzio_server' object

Usage

quetzio_server$new(
  source_method,
  source_yaml = NULL,
  source_yaml_default = NULL,
  source_gsheet_id = NULL,
  source_gsheet_sheetname = NULL,
  source_object = NULL,
  output_gsheet = FALSE,
  output_gsheet_id = NULL,
  output_gsheet_sheetname = NULL,
  desc_yaml = NULL,
  desc_gsheet_id = NULL,
  desc_gsheet_sheetname = NULL,
  desc_object = NULL,
  module_id = NULL,
  div_id = NULL,
  custom_css = NULL,
  button_labels = c("Submit", "Cannot submit", "Submitted!", "Error occured!"),
  render_ui = TRUE,
  link_id = NULL
)

Arguments

source_method

character string specifying in what form the source config file will be provided. Can be either 'gsheet', 'yaml' or 'raw'. Necessity of other arguments is dependent on this choice. For more info see 'details'

source_yaml

path to the source yaml file

source_yaml_default

path to the optional default options for items generated with 'source_yaml' file. Only when source_method == 'yaml'.

source_gsheet_id

id of the source googlesheet file

source_gsheet_sheetname

name of the source spreadsheet

source_object

object of class list (similiar in structure to 'yaml' source) or data.frame (similiar in structure to 'googlesheet' source) to be the source of questions. You can create a sample data.frame with create_survey_source(). Needed when source_method == 'raw'

output_gsheet

logical: do you wish to save the answers automatically to the googlesheet. If TRUE, the 'output_gsheet_id' and 'output_gsheet_sheetname' arguments need to be specified. Defaults to FALSE

output_gsheet_id

id of the output googlesheet file. If not specified, the same googlesheet as for 'source' will be used

output_gsheet_sheetname

name of the output spreadsheet

desc_yaml

path to the optional instruction and item descriptions.

desc_gsheet_id

id of the googlesheet to provide optional instruction and item descriptions. Defaults to 'source_gsheet_id', if not provided.

desc_gsheet_sheetname

name of source for optional instruction and item descriptions.

desc_object

object of class list or data.frame to be the source of optional instruction and item descriptions.

module_id

character string with unique id for the module. If not specified, it will be automatically generated

div_id

character string with unique id for the created div. If not specified, it will be set to 'form'

custom_css

custom css for classes 'mandatory star' and 'invalid_input'. If not specified, default look will be used:

invalid_input = "outline: red; outline-style: dashed; outline-offset: 10px;"

mandatory_star = "color: red;"

button_labels

character vector of length four with labels for submit button i all states. Defaults to: c('Submit', 'Cannot submit', 'Submitted!', 'Error occured!')

render_ui

logical indicating if the UI for questionnaire should be rendered

link_id

character specifying the 'link_id' of the 'quetzio_link_server' object, modifying its namespace. Only used internally, if the questionnaire is part of linked server. Don't set it manually.

Details

Currently, there are multiple methods both for source, which will generate the inputs, and for output. Mandatory arguments change depending of your choices:

for source:

source_method == 'yaml': 'source_yaml'

source_method == 'gsheet': 'source_gsheet_id,' 'source_gsheet_sheetname'

source_method == 'raw': 'source_object'

for output:

output_gsheet == TRUE: 'output_gsheet_id' (if other than 'source_gsheet_id') and 'output_gsheet_sheetname'

There are also some optional functionalities, that can be used with sources.

optional instructions and item descriptions - they are generated only if one of the following is provided:

desc_yaml: rendering from YAML file

desc_gsheet_sheetname: rendering from googlesheet. If the source_method isn't gsheet or the 'googlesheet_id' containing description is different from source, the desc_gsheet_id need to be provided too

desc_object: rendering from R object of classes 'data.frame' or 'list'

optional default configuration - it is used for a type of shinyInput. Need to provide source_yaml_default - there are no other methods ATM.

Returns

the 'quetzio_server' serverModule

Method get_answers_df()

method to get preprocessed answers in the form of dataframe (only if all of the questionnaires are done)

Usage

quetzio_server$get_answers_df()

Returns

data.frame

Method update_labels()

Method to update labels on the change in reactive

Usage

quetzio_server$update_labels(
  trigger,
  source_method,
  source_yaml = NULL,
  source_gsheet_id = NULL,
  source_gsheet_sheetname = NULL,
  source_object = NULL
)

Arguments

trigger

reactive which will trigger the update. It needs to take values linked to the changes in the source

source_method

character string specifying in what form the source config file will be provided. Can be either 'gsheet', 'yaml' or 'raw'. Necessity of other arguments is dependent on this choice

source_yaml

path to the source yaml file

source_gsheet_id

id of the source googlesheet file

source_gsheet_sheetname

name of the source spreadsheet

source_object

object of class list (similiar in structure to 'yaml' source) or data.frame (similiar in structure to 'googlesheet' source) to be the source of questions. You can create a sample data.frame with create_survey_source(). Needed when source_method == 'raw'

Method clone()

The objects of this class are cloneable with this method.

Usage

quetzio_server$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.